==> 4. Inspecting Datatypes.ipynb <==
# Inspecting Data Types
Use the space below to explore `data_08_v2.csv` and `data_18_v2.csv` to answer the quiz questions below regarding datatypes. You should've created these data files in the previous section: *Filter, Drop Nulls, Dedupe*.

# importing libraries
import pandas as pd

# loading the datasets
df_08 = pd.read_csv("data_08_v2.csv")
df_18 = pd.read_csv("data_18_v2.csv")

df_08.info()

df_18.info()

'''Which of the following changes could be made to make the 
cyl column in both datasets clear and consistent?'''
print(df_08.cyl.describe())
print(df_18.cyl.describe())

'''Which of the following changes could be made to make the air_pollution_score column in both 
datasets consistent?'''
print(df_08.air_pollution_score.value_counts())
print(df_18.air_pollution_score.value_counts())

df_08[['city_mpg', 'hwy_mpg', 'cmb_mpg']].head()

# needs to be converted from string to int
df_18[['city_mpg', 'hwy_mpg', 'cmb_mpg']].head()

# # needs to be converted from string to int
df_08.greenhouse_gas_score.head()

'''Which changes need to be made to make the greenhouse_gas_score 
columns in both datasets consistent?'''
df_18.greenhouse_gas_score.head()
